+++ quickbooks/auth.py
@@ -47,103 +47,6 @@ class AuthSessionManager(object):
         return self.session
 
 
-class Oauth1SessionManager(AuthSessionManager):
-    oauth_version = 1.0
-    request_token_url = "https://oauth.intuit.com/oauth/v1/get_request_token"
-    access_token_url = "https://oauth.intuit.com/oauth/v1/get_access_token"
-    authorize_url = "https://appcenter.intuit.com/Connect/Begin"
-
-    request_token = ''
-    request_token_secret = ''
-
-    def __init__(self, **kwargs):
-        if 'consumer_key' in kwargs:
-            self.consumer_key = kwargs['consumer_key']
-
-        if 'consumer_secret' in kwargs:
-            self.consumer_secret = kwargs['consumer_secret']
-
-        if 'access_token' in kwargs:
-            self.access_token = kwargs['access_token']
-
-        if 'access_token_secret' in kwargs:
-            self.access_token_secret = kwargs['access_token_secret']
-
-    def start_session(self):
-        if not self.started:
-            if self.consumer_key == '':
-                raise QuickbooksException("Consumer Key missing. Cannot create session.")
-
-            if self.consumer_secret == '':
-                raise QuickbooksException("Consumer Secret missing. Cannot create session.")
-
-            if self.access_token == '':
-                raise QuickbooksException("Access Token missing. Cannot create session.")
-
-            if self.access_token_secret == '':
-                raise QuickbooksException("Access Token Secret missing. Cannot create session.")
-
-            self.session = OAuth1Session(
-                self.consumer_key,
-                self.consumer_secret,
-                self.access_token,
-                self.access_token_secret,
-            )
-
-            self.started = True
-
-        return self.session
-
-    def get_authorize_url(self, callback_url):
-        """
-        Returns the Authorize URL as returned by QB, and specified by OAuth 1.0a.
-        :return URI:
-        """
-        self.authorize_url = self.authorize_url[:self.authorize_url.find('?')] \
-            if '?' in self.authorize_url else self.authorize_url
-
-        qb_service = OAuth1Service(
-            consumer_key=self.consumer_key,
-            consumer_secret=self.consumer_secret,
-            request_token_url=self.request_token_url,
-            access_token_url=self.access_token_url,
-            authorize_url=self.authorize_url,
-        )
-
-        response = qb_service.get_raw_request_token(
-            params={'oauth_callback': callback_url})
-
-        oauth_resp = dict(parse_qsl(response.text))
-
-        self.request_token = oauth_resp['oauth_token']
-        self.request_token_secret = oauth_resp['oauth_token_secret']
-
-        return qb_service.get_authorize_url(self.request_token)
-
-    def get_access_tokens(self, oauth_verifier):
-        """
-        Wrapper around get_auth_session, returns session, and sets access_token and
-        access_token_secret on the QB Object.
-        :param oauth_verifier: the oauth_verifier as specified by OAuth 1.0a
-        """
-        qb_service = OAuth1Service(
-            consumer_key=self.consumer_key,
-            consumer_secret=self.consumer_secret,
-            request_token_url=self.request_token_url,
-            access_token_url=self.access_token_url,
-            authorize_url=self.authorize_url,
-        )
-
-        session = qb_service.get_auth_session(
-            self.request_token,
-            self.request_token_secret,
-            data={'oauth_verifier': oauth_verifier})
-
-        self.access_token = session.access_token
-        self.access_token_secret = session.access_token_secret
-        return session
-
-
 class Oauth2SessionManager(AuthSessionManager):
     oauth_version = 2.0
     access_token_url = "https://oauth.platform.intuit.com/oauth2/v1/tokens/bearer"
@@ -171,6 +74,9 @@ class Oauth2SessionManager(AuthSessionManager):
 
         if 'base_url' in kwargs:
             self.base_url = kwargs['base_url']
+
+        if 'refresh_token' in kwargs:
+            self.refresh_token = kwargs['refresh_token']
 
     def start_session(self):
         if not self.started:
@@ -219,25 +125,7 @@ class Oauth2SessionManager(AuthSessionManager):
 
         return url
 
-    def get_access_tokens(self, auth_code):
-        headers = {
-            'Accept': 'application/json',
-            'content-type': 'application/x-www-form-urlencoded',
-            'Authorization': self.get_auth_header()
-        }
-
-        payload = {
-            'code': auth_code,
-            'redirect_uri': self.base_url,
-            'grant_type': 'authorization_code'
-        }
-
-        r = requests.post(self.access_token_url, data=payload, headers=headers)
-        if r.status_code != 200:
-            return r.text
-
-        bearer_raw = json.loads(r.text)
-
+    def update_tokens(self, bearer_raw):
         self.x_refresh_token_expires_in = bearer_raw['x_refresh_token_expires_in']
         self.access_token = bearer_raw['access_token']
         self.token_type = bearer_raw['token_type']
@@ -254,3 +142,35 @@ class Oauth2SessionManager(AuthSessionManager):
             auth_header = base64.b64encode(bytes(self.client_id + ':' + self.client_secret, 'utf-8')).decode()
 
         return 'Basic ' + auth_header
+
+    def token_request(self, payload, return_result=False):
+        headers = {
+            'Accept': 'application/json',
+            'content-type': 'application/x-www-form-urlencoded',
+            'Authorization': self.get_auth_header()
+        }
+        r = requests.post(self.access_token_url, data=payload, headers=headers)
+        if r.status_code != 200:
+            return r.text
+
+        bearer_raw = json.loads(r.text)
+
+        self.update_tokens(bearer_raw)
+
+        return bearer_raw if return_result else None
+
+    def get_access_tokens(self, auth_code, return_result=False):
+        payload = {
+            'code': auth_code,
+            'redirect_uri': self.base_url,
+            'grant_type': 'authorization_code'
+        }
+        return self.token_request(payload, return_result=return_result)
+
+    def refresh_access_tokens(self, refresh_token=None, return_result=False):
+        payload = {
+            'refresh_token': refresh_token or self.refresh_token,
+            'grant_type': 'refresh_token'
+        }
+        return self.token_request(payload, return_result=return_result)
+