+++ quickbooks/client.py
@@ -1,18 +1,24 @@
 try:  # Python 3
     import http.client as httplib
     from urllib.parse import parse_qsl
+    from functools import partial
+    to_bytes = lambda value, *args, **kwargs: bytes(value, "utf-8", *args, **kwargs)
 except ImportError:  # Python 2
     import httplib
     from urlparse import parse_qsl
+    to_bytes = str
 
 import textwrap
+import codecs
 import json
 
-from .exceptions import QuickbooksException, SevereException, AuthorizationException
+from . import exceptions
 import base64
+import hashlib
+import hmac
 
 try:
-    from rauth import OAuth1Session, OAuth1Service
+    from rauth import OAuth1Session, OAuth1Service, OAuth2Session
 except ImportError:
     print("Please import Rauth:\n\n")
     print("http://rauth.readthedocs.org/en/latest/\n")
@@ -22,9 +28,10 @@ except ImportError:
 class QuickBooks(object):
     company_id = 0
     session = None
-    session_manager = None
+    auth_client = None
     sandbox = False
     minorversion = None
+    verifier_token = None
 
     sandbox_api_url_v3 = "https://sandbox-quickbooks.api.intuit.com/v3"
     api_url_v3 = "https://quickbooks.api.intuit.com/v3"
@@ -35,11 +42,11 @@ class QuickBooks(object):
 
     _BUSINESS_OBJECTS = [
         "Account", "Attachable", "Bill", "BillPayment",
-        "Class", "CreditMemo", "Customer",
+        "Class", "CreditMemo", "Customer", "CompanyCurrency",
         "Department", "Deposit", "Employee", "Estimate", "Invoice",
         "Item", "JournalEntry", "Payment", "PaymentMethod",
         "Purchase", "PurchaseOrder", "RefundReceipt",
-        "SalesReceipt", "TaxCode", "TaxService/Taxcode", "TaxRate", "Term",
+        "SalesReceipt", "TaxAgency", "TaxCode", "TaxService/Taxcode", "TaxRate", "Term",
         "TimeActivity", "Transfer", "Vendor", "VendorCredit"
     ]
 
@@ -57,6 +64,13 @@ class QuickBooks(object):
         else:
             instance = object.__new__(cls)
 
+        if 'refresh_token' in kwargs:
+            instance.refresh_token = kwargs['refresh_token']
+
+        if 'auth_client' in kwargs:
+            instance.auth_client = kwargs['auth_client']
+            instance._start_session()
+
         if 'company_id' in kwargs:
             instance.company_id = kwargs['company_id']
 
@@ -66,10 +80,20 @@ class QuickBooks(object):
         if 'minorversion' in kwargs:
             instance.minorversion = kwargs['minorversion']
 
-        if 'session_manager' in kwargs:
-            instance.session_manager = kwargs['session_manager']
+        if 'verifier_token' in kwargs:
+            instance.verifier_token = kwargs.get('verifier_token')
 
         return instance
+
+    def _start_session(self):
+        if self.auth_client.access_token is None:
+            self.auth_client.refresh(refresh_token=self.refresh_token)
+
+        self.session = OAuth2Session(
+            client_id=self.auth_client.client_id,
+            client_secret=self.auth_client.client_secret,
+            access_token=self.auth_client.access_token,
+        )
 
     @classmethod
     def get_instance(cls):
@@ -100,10 +124,19 @@ class QuickBooks(object):
         else:
             return self.api_url_v3
 
+    def validate_webhook_signature(self, request_body, signature, verifier_token=None):
+        hmac_verifier_token_hash = hmac.new(
+            to_bytes(verifier_token or self.verifier_token),
+            request_body.encode('utf-8'),
+            hashlib.sha256
+        ).digest()
+        decoded_hex_signature = base64.b64decode(signature)
+        return hmac_verifier_token_hash == decoded_hex_signature
+
     def get_current_user(self):
         """Get data from the current user endpoint"""
         url = self.current_user_url
-        result = self.make_request("GET", url)
+        result = self.get(url)
         return result
 
     def get_report(self, report_type, qs=None):
@@ -112,35 +145,33 @@ class QuickBooks(object):
             qs = {}
 
         url = self.api_url + "/company/{0}/reports/{1}".format(self.company_id, report_type)
-        result = self.make_request("GET", url, params=qs)
+        result = self.get(url, params=qs)
         return result
 
-    # TODO: is disconnect url the same for OAuth 1 and OAuth 2?
     def disconnect_account(self):
         """
         Disconnect current account from the application
         :return:
         """
         url = self.disconnect_url
-        result = self.make_request("GET", url)
+        result = self.get(url)
         return result
 
     def change_data_capture(self, entity_string, changed_since):
-        url = self.api_url + "/company/{0}/cdc".format(self.company_id)
+        url = "{0}/company/{1}/cdc".format(self.api_url, self.company_id)
 
         params = {"entities": entity_string, "changedSince": changed_since}
 
-        result = self.make_request("GET", url, params=params)
+        result = self.get(url, params=params)
         return result
 
-    # TODO: is reconnect url the same for OAuth 1 and OAuth 2?
     def reconnect_account(self):
         """
         Reconnect current account by refreshing OAuth access tokens
         :return:
         """
         url = self.reconnect_url
-        result = self.make_request("GET", url)
+        result = self.get(url)
         return result
 
     def make_request(self, request_type, url, request_body=None, content_type='application/json',
@@ -201,38 +232,39 @@ class QuickBooks(object):
         req = self.process_request(request_type, url, headers=headers, params=params, data=request_body)
 
         if req.status_code == httplib.UNAUTHORIZED:
-            raise AuthorizationException("Application authentication failed", detail=req.text)
+            raise exceptions.AuthorizationException("Application authentication failed", detail=req.text)
 
         try:
             result = req.json()
         except:
-            raise QuickbooksException("Error reading json response: {0}".format(req.text), 10000)
+            raise exceptions.QuickbooksException("Error reading json response: {0}".format(req.text), 10000)
 
         if "Fault" in result:
             self.handle_exceptions(result["Fault"])
         elif not req.status_code == httplib.OK:
-            raise QuickbooksException("Error returned with status code '{0}': {1}".format(
+            raise exceptions.QuickbooksException("Error returned with status code '{0}': {1}".format(
                 req.status_code, req.text), 10000)
         else:
             return result
 
+    def get(self, *args, **kwargs):
+        return self.make_request("GET", *args, **kwargs)
+
+    def post(self, *args, **kwargs):
+        return self.make_request("POST", *args, **kwargs)
+
     def process_request(self, request_type, url, headers="", params="", data=""):
-        if self.session_manager is None:
-            raise QuickbooksException('No session manager')
+        if self.session is None:
+            raise exceptions.QuickbooksException('No session manager')
 
-        if self.session_manager.oauth_version == 2.0:
-            headers.update({'Authorization': 'Bearer ' + self.session_manager.access_token})
-            return self.session_manager.get_session().request(
-                request_type, url, headers=headers, params=params, data=data)
+        headers.update({'Authorization': 'Bearer ' + self.session.access_token})
 
-        else:
-            return self.session_manager.get_session().request(
-                request_type, url, True, self.company_id,
-                headers=headers, params=params, data=data)
+        return self.session.request(
+            request_type, url, headers=headers, params=params, data=data)
 
     def get_single_object(self, qbbo, pk):
-        url = self.api_url + "/company/{0}/{1}/{2}/".format(self.company_id, qbbo.lower(), pk)
-        result = self.make_request("GET", url, {})
+        url = "{0}/company/{1}/{2}/{3}/".format(self.api_url, self.company_id, qbbo.lower(), pk)
+        result = self.get(url, {})
 
         return result
 
@@ -250,22 +282,32 @@ class QuickBooks(object):
             if "code" in error:
                 code = int(error["code"])
 
-            if code >= 10000:
-                raise SevereException(message, code, detail)
+            if code > 0 and code <= 499:
+                raise exceptions.AuthorizationException(message, code, detail)
+            elif code >= 500 and code <= 599:
+                raise exceptions.UnsupportedException(message, code, detail)
+            elif code >= 600 and code <= 1999:
+                if code == 610:
+                    raise exceptions.ObjectNotFoundException(message, code, detail)
+                raise exceptions.GeneralException(message, code, detail)
+            elif code >= 2000 and code <= 4999:
+                raise exceptions.ValidationException(message, code, detail)
+            elif code >= 10000:
+                raise exceptions.SevereException(message, code, detail)
             else:
-                raise QuickbooksException(message, code, detail)
+                raise exceptions.QuickbooksException(message, code, detail)
 
     def create_object(self, qbbo, request_body, _file_path=None):
         self.isvalid_object_name(qbbo)
 
-        url = self.api_url + "/company/{0}/{1}".format(self.company_id, qbbo.lower())
-        results = self.make_request("POST", url, request_body, file_path=_file_path)
+        url = "{0}/company/{1}/{2}".format(self.api_url, self.company_id, qbbo.lower())
+        results = self.post(url, request_body, file_path=_file_path)
 
         return results
 
     def query(self, select):
-        url = self.api_url + "/company/{0}/query".format(self.company_id)
-        result = self.make_request("POST", url, select, content_type='application/text')
+        url = "{0}/company/{1}/query".format(self.api_url, self.company_id)
+        result = self.post(url, select, content_type='application/text')
 
         return result
 
@@ -276,35 +318,35 @@ class QuickBooks(object):
         return True
 
     def update_object(self, qbbo, request_body, _file_path=None):
-        url = self.api_url + "/company/{0}/{1}".format(self.company_id, qbbo.lower())
-        result = self.make_request("POST", url, request_body, file_path=_file_path)
+        url = "{0}/company/{1}/{2}".format(self.api_url, self.company_id,  qbbo.lower())
+        result = self.post(url, request_body, file_path=_file_path)
 
         return result
 
     def delete_object(self, qbbo, request_body, _file_path=None):
-        url = self.api_url + "/company/{0}/{1}".format(self.company_id, qbbo.lower())
-        result = self.make_request("POST", url, request_body, params={'operation': 'delete'}, file_path=_file_path)
+        url = "{0}/company/{1}/{2}".format(self.api_url, self.company_id, qbbo.lower())
+        result = self.post(url, request_body, params={'operation': 'delete'}, file_path=_file_path)
 
         return result
 
     def batch_operation(self, request_body):
-        url = self.api_url + "/company/{0}/batch".format(self.company_id)
-        results = self.make_request("POST", url, request_body)
+        url = "{0}/company/{1}/batch".format(self.api_url, self.company_id)
+        results = self.post(url, request_body)
 
         return results
 
-    def misc_operation(self, end_point, request_body):
-        url = self.api_url + "/company/{0}/{1}".format(self.company_id, end_point)
-        results = self.make_request("POST", url, request_body)
+    def misc_operation(self, end_point, request_body, content_type='application/json'):
+        url = "{0}/company/{1}/{2}".format(self.api_url, self.company_id, end_point)
+        results = self.post(url, request_body, content_type)
 
         return results
 
     def download_pdf(self, qbbo, item_id):
-        if self.session_manager is None:
-            raise QuickbooksException('No session manager')
+        if self.session is None:
+            raise exceptions.QuickbooksException('No session')
 
-        url = self.api_url + "/company/{0}/{1}/{2}/pdf".format(
-            self.company_id, qbbo.lower(), item_id)
+        url = "{0}/company/{1}/{2}/{3}/pdf".format(
+            self.api_url, self.company_id, qbbo.lower(), item_id)
 
         headers = {
             'Content-Type': 'application/pdf',
@@ -315,10 +357,15 @@ class QuickBooks(object):
         response = self.process_request("GET", url, headers=headers)
 
         if response.status_code != httplib.OK:
+
+            if response.status_code == httplib.UNAUTHORIZED:
+                # Note that auth errors have different result structure which can't be parsed by handle_exceptions()
+                raise exceptions.AuthorizationException("Application authentication failed", detail=response.text)
+
             try:
                 result = response.json()
             except:
-                raise QuickbooksException("Error reading json response: {0}".format(response.text), 10000)
+                raise exceptions.QuickbooksException("Error reading json response: {0}".format(response.text), 10000)
 
             self.handle_exceptions(result["Fault"])
         else: