+++ quickbooks/objects/detailline.py
@@ -44,6 +44,7 @@ class DiscountLineDetail(QuickbooksBaseObject):
         "Discount": DiscountOverride,
         "ClassRef": Ref,
         "TaxCodeRef": Ref,
+        "DiscountAccountRef": Ref
     }
 
     def __init__(self):
@@ -52,6 +53,8 @@ class DiscountLineDetail(QuickbooksBaseObject):
         self.Discount = None
         self.ClassRef = None
         self.TaxCodeRef = None
+        self.PercentBased = False
+        self.DiscountPercent = 0
 
 
 class DiscountLine(DetailLine):
@@ -220,7 +223,7 @@ class ItemBasedExpenseLineDetail(QuickbooksBaseObject):
 
     def __init__(self):
         super(ItemBasedExpenseLineDetail, self).__init__()
-        self.BillableStatus = ""
+        self.BillableStatus = None
         self.UnitPrice = 0
         self.TaxInclusiveAmt = 0
         self.Qty = 0