+++ tests/integration/test_department.py
@@ -1,28 +1,12 @@
-import os
-import unittest
 from datetime import datetime
 
-from quickbooks.auth import Oauth1SessionManager
 from quickbooks.objects.department import Department
+from tests.integration.test_base import QuickbooksTestCase
 
-from quickbooks import QuickBooks
 
-
-class DepartmentTest(unittest.TestCase):
+class DepartmentTest(QuickbooksTestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
+        super(DepartmentTest, self).setUp()
 
         self.name = "Test Department {0}".format(datetime.now().strftime('%d%H%M%S'))