+++ tests/integration/test_deposit.py
@@ -1,30 +1,9 @@
-import os
-import unittest
-
-from quickbooks.auth import Oauth1SessionManager
 from quickbooks.objects.account import Account
-
 from quickbooks.objects.deposit import Deposit, DepositLine, DepositLineDetail
+from tests.integration.test_base import QuickbooksTestCase
 
-from quickbooks import QuickBooks
 
-
-class DepositTest(unittest.TestCase):
-    def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
-
+class DepositTest(QuickbooksTestCase):
     def test_create(self):
         deposit = Deposit()
         account = Account.filter(AccountType="Bank", max_results=2, qb=self.qb_client)