+++ tests/integration/test_invoice.py
@@ -1,31 +1,12 @@
-import os
-import unittest
-
-from quickbooks.auth import Oauth1SessionManager
-from quickbooks.client import QuickBooks
 from quickbooks.objects.base import CustomerMemo
 from quickbooks.objects.customer import Customer
 from quickbooks.objects.detailline import SalesItemLine, SalesItemLineDetail
 from quickbooks.objects.invoice import Invoice
 from quickbooks.objects.item import Item
+from tests.integration.test_base import QuickbooksTestCase
 
 
-class InvoiceTest(unittest.TestCase):
-    def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
-
+class InvoiceTest(QuickbooksTestCase):
     def test_create(self):
         invoice = Invoice()