+++ tests/integration/test_vendor.py
@@ -1,28 +1,13 @@
-import os
-import unittest
 from datetime import datetime
 
-from quickbooks.auth import Oauth1SessionManager
-from quickbooks.client import QuickBooks
 from quickbooks.objects.base import Address, PhoneNumber, EmailAddress, WebAddress
 from quickbooks.objects.vendor import Vendor
+from tests.integration.test_base import QuickbooksTestCase
 
 
-class VendorTest(unittest.TestCase):
+class VendorTest(QuickbooksTestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
+        super(VendorTest, self).setUp()
 
         self.account_number = datetime.now().strftime('%d%H%M')
         self.name = "Test Vendor {0}".format(self.account_number)